+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## @sgx4u/date-time-utils - Changelog
+# @sgx4u/date-time-utils - Changelog
 
 ## 2.1.0